pyproject.toml
[project]
name = "resumeanalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bert-score>=0.3.13",
    "chromadb>=1.0.15",
    "datasets>=4.0.0",
    "diversity>=0.2.2",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.3.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.0.0",
    "spacy>=3.8.7",
    "tiktoken>=0.9.0",
]
